Tests: move bindNullEngine to helper module
When it is imported from CryptoClientTest, it causes Jest to parse the
tests in that file multiple times.  Moving it to a standalone, non-test
module avoids that.
