+++ test/encryption/RoomTrackerTest.ts
@@ -2,7 +2,7 @@ import * as simple from "simple-mock";
 
 import { EncryptionEventContent, MatrixClient, RoomEncryptionAlgorithm, RoomTracker } from "../../src";
 import { createTestClient, testCryptoStores, TEST_DEVICE_ID } from "../TestUtils";
-import { bindNullEngine } from "./CryptoClientTest";
+import { bindNullEngine } from "./CryptoTestUtils";
 
 function prepareQueueSpies(
     client: MatrixClient,
